Move everything exclusively to GraphQL
