--- 3_ecosystem/static/index.html
@@ -1,38 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>Step 3</title>
-</head>
-<body>
-    <form id="loginForm" method="POST">
-        <input type="text" placeholder="name" name="name">
-        <input type="text" placeholder="password" name="password">
-        <input type="submit" value="login">
-    </form>
-    <button id="getLoginData">get login data</button>
-    <script>
-        document.getElementById('loginForm').addEventListener('submit', async (event) => {
-            event.preventDefault();
-            let formData = new FormData(document.getElementById('loginForm'));
-            var object = {};
-            formData.forEach((value, key) => {object[key] = value});
-            console.log(object);
-            const response = await fetch('/login', {
-                method: 'POST',
-                headers: {
-                    'Content-Type': 'application/json',
-                },
-                body: JSON.stringify(object),
-            });
-            alert(await response.text())
-        });
-
-        document.getElementById('getLoginData').addEventListener('click', async (event) => {
-            const response = await fetch('/info');
-            alert(await response.text())
-        })
-    </script>
-</body>
-</html>
